+++ ouroboros-network/src/Ouroboros/Network/TxSubmission/Inbound/V1.hs
@@ -1,5 +1,4 @@
 {-# LANGUAGE BangPatterns        #-}
-{-# LANGUAGE CPP                 #-}
 {-# LANGUAGE DataKinds           #-}
 {-# LANGUAGE DeriveGeneric       #-}
 {-# LANGUAGE GADTs               #-}
@@ -9,8 +8,9 @@
 
 {-# OPTIONS_GHC -Wno-partial-fields #-}
 
-module Ouroboros.Network.TxSubmission.Inbound
+module Ouroboros.Network.TxSubmission.Inbound.V1
   ( txSubmissionInbound
+  , TxSubmissionInitDelay (..)
   , TxSubmissionMempoolWriter (..)
   , TraceTxSubmissionInbound (..)
   , TxSubmissionProtocolError (..)
@@ -21,7 +21,6 @@ import Data.Foldable as Foldable (foldl', toList)
 import Data.List.NonEmpty qualified as NonEmpty
 import Data.Map.Strict (Map)
 import Data.Map.Strict qualified as Map
-import Data.Maybe
 import Data.Sequence.Strict (StrictSeq)
 import Data.Sequence.Strict qualified as Seq
 import Data.Set qualified as Set
@@ -36,71 +35,20 @@ import Control.Exception (assert)
 import Control.Monad (unless)
 import Control.Monad.Class.MonadSTM
 import Control.Monad.Class.MonadThrow
+import Control.Monad.Class.MonadTime.SI
 import Control.Monad.Class.MonadTimer.SI
 import Control.Tracer (Tracer, traceWith)
 
 import Network.TypedProtocol.Core (N, Nat (..), natToInt)
 
 import Ouroboros.Network.NodeToNode.Version (NodeToNodeVersion)
-import Ouroboros.Network.Protocol.Limits
 import Ouroboros.Network.Protocol.TxSubmission2.Server
 import Ouroboros.Network.Protocol.TxSubmission2.Type
+import Ouroboros.Network.TxSubmission.Inbound.V2.Types (ProcessedTxCount (..),
+           TraceTxSubmissionInbound (..), TxSubmissionInitDelay (..),
+           TxSubmissionMempoolWriter (..), TxSubmissionProtocolError (..))
 import Ouroboros.Network.TxSubmission.Mempool.Reader (MempoolSnapshot (..),
            TxSubmissionMempoolReader (..))
-
--- | The consensus layer functionality that the inbound side of the tx
--- submission logic requires.
---
--- This is provided to the tx submission logic by the consensus layer.
---
-data TxSubmissionMempoolWriter txid tx idx m =
-     TxSubmissionMempoolWriter {
-
-       -- | Compute the transaction id from a transaction.
-       --
-       -- This is used in the protocol handler to verify a full transaction
-       -- matches a previously given transaction id.
-       --
-       txId          :: tx -> txid,
-
-       -- | Supply a batch of transactions to the mempool. They are either
-       -- accepted or rejected individually, but in the order supplied.
-       --
-       -- The 'txid's of all transactions that were added successfully are
-       -- returned.
-       mempoolAddTxs :: [tx] -> m [txid]
-    }
-
-data ProcessedTxCount = ProcessedTxCount {
-      -- | Just accepted this many transactions.
-      ptxcAccepted :: Int
-      -- | Just rejected this many transactions.
-    , ptxcRejected :: Int
-    }
-  deriving (Eq, Show)
-
-data TraceTxSubmissionInbound txid tx =
-    -- | Number of transactions just about to be inserted.
-    TraceTxSubmissionCollected Int
-    -- | Just processed transaction pass/fail breakdown.
-  | TraceTxSubmissionProcessed ProcessedTxCount
-    -- | Server received 'MsgDone'
-  | TraceTxInboundTerminated
-  | TraceTxInboundCanRequestMoreTxs Int
-  | TraceTxInboundCannotRequestMoreTxs Int
-  deriving (Eq, Show)
-
-data TxSubmissionProtocolError =
-       ProtocolErrorTxNotRequested
-     | ProtocolErrorTxIdsNotRequested
-  deriving Show
-
-instance Exception TxSubmissionProtocolError where
-  displayException ProtocolErrorTxNotRequested =
-      "The peer replied with a transaction we did not ask for."
-  displayException ProtocolErrorTxIdsNotRequested =
-      "The peer replied with more txids than we asked for."
-
 
 -- | Information maintained internally in the 'txSubmissionInbound' server
 -- implementation.
@@ -183,18 +131,17 @@ txSubmissionInbound
      , MonadDelay m
      )
   => Tracer m (TraceTxSubmissionInbound txid tx)
+  -> TxSubmissionInitDelay
   -> NumTxIdsToAck  -- ^ Maximum number of unacknowledged txids allowed
   -> TxSubmissionMempoolReader txid tx idx m
   -> TxSubmissionMempoolWriter txid tx idx m
   -> NodeToNodeVersion
   -> TxSubmissionServerPipelined txid tx m ()
-txSubmissionInbound tracer (NumTxIdsToAck maxUnacked) mpReader mpWriter _version =
+txSubmissionInbound tracer initDelay (NumTxIdsToAck maxUnacked) mpReader mpWriter _version =
     TxSubmissionServerPipelined $ do
-#ifdef TXSUBMISSION_DELAY
-      -- make the client linger before asking for tx's and expending
-      -- our resources as well, as he may disconnect for some reason
-      threadDelay (fromMaybe (-1) longWait)
-#endif
+      case initDelay of
+        TxSubmissionInitDelay delay -> threadDelay delay
+        NoTxSubmissionInitDelay     -> return ()
       continueWithStateM (serverIdle Zero) initialServerState
   where
     -- TODO #1656: replace these fixed limits by policies based on
@@ -262,7 +209,7 @@ txSubmissionInbound tracer (NumTxIdsToAck maxUnacked) mpReader mpWriter _version
             --
             traceWith tracer (TraceTxInboundCanRequestMoreTxs (natToInt n))
             pure $ CollectPipelined
-              (Just (continueWithState (serverReqTxs (Succ n')) st))
+              (Just (pure $ continueWithState (serverReqTxs (Succ n')) st))
               (collectAndContinueWithState (handleReply n') st)
 
           else do
@@ -317,7 +264,7 @@ txSubmissionInbound tracer (NumTxIdsToAck maxUnacked) mpReader mpWriter _version
             txsMap = Map.fromList [ (txId tx, tx) | tx <- txs ]
 
             txidsReceived  = Map.keysSet txsMap
-            txidsRequested = Set.fromList txids
+            txidsRequested = Map.keysSet txids
 
         unless (txidsReceived `Set.isSubsetOf` txidsRequested) $
           throwIO ProtocolErrorTxNotRequested
@@ -361,17 +308,21 @@ txSubmissionInbound tracer (NumTxIdsToAck maxUnacked) mpReader mpWriter _version
             bufferedTxs3 = forceElemsToWHNF $ bufferedTxs2 <>
                                Map.fromList (zip live (repeat Nothing))
 
-        let !collected = length txs
         traceWith tracer $
-          TraceTxSubmissionCollected collected
+          TraceTxSubmissionCollected (txId `map` txs)
 
+        !start <- getMonotonicTime
         txidsAccepted <- mempoolAddTxs txsReady
-
+        !end <- getMonotonicTime
+        let duration = diffTime end start
+        traceWith tracer $
+          TraceTxInboundAddedToMempool txidsAccepted duration
         let !accepted = length txidsAccepted
 
         traceWith tracer $ TraceTxSubmissionProcessed ProcessedTxCount {
             ptxcAccepted = accepted
-          , ptxcRejected = collected - accepted
+          , ptxcRejected = length txs - accepted
+          , ptxcScore    = 0 -- This implementation does not track score
           }
 
         continueWithStateM (serverIdle n) st {
@@ -464,7 +415,7 @@ txSubmissionInbound tracer (NumTxIdsToAck maxUnacked) mpReader mpWriter _version
               Map.splitAt (fromIntegral maxTxToRequest) (availableTxids st)
 
         SendMsgRequestTxsPipelined
-          (Map.keys txsToRequest)
+          txsToRequest
           (continueWithStateM (serverReqTxIds (Succ n)) st {
              availableTxids = availableTxids'
            })