+++ ouroboros-network/testlib/Test/Ouroboros/Network/Diffusion/Node/Kernel.hs
@@ -13,6 +13,8 @@ module Test.Ouroboros.Network.Diffusion.Node.Kernel
   , encodeNtNAddr
   , decodeNtNAddr
   , ntnAddrToRelayAccessPoint
+  , ppNtNAddr
+  , ppNtNConnId
   , NtNVersion
   , NtNVersionData (..)
   , NtCAddr
@@ -36,6 +38,7 @@ import Control.Concurrent.Class.MonadSTM.Strict
 import Control.DeepSeq (NFData (..))
 import Control.Monad (replicateM, when)
 import Control.Monad.Class.MonadAsync
+import Control.Monad.Class.MonadFork
 import Control.Monad.Class.MonadThrow
 import Control.Monad.Class.MonadTime.SI
 import Control.Monad.Class.MonadTimer.SI
@@ -52,7 +55,7 @@ import Data.Monoid.Synchronisation
 import Data.Void (Void)
 import GHC.Generics (Generic)
 import Numeric.Natural (Natural)
-import System.Random (SplitGen, StdGen, randomR, splitGen)
+import System.Random (SplitGen, StdGen, mkStdGen, random, randomR, splitGen)
 
 import Network.Socket (PortNumber)
 
@@ -60,12 +63,17 @@ import Ouroboros.Network.AnchoredFragment (Anchor (..))
 import Ouroboros.Network.Block (HasFullHeader, SlotNo)
 import Ouroboros.Network.Block qualified as Block
 import Ouroboros.Network.BlockFetch
+import Ouroboros.Network.ConnectionId (ConnectionId (..))
 import Ouroboros.Network.Handshake.Acceptable (Accept (..), Acceptable (..))
 import Ouroboros.Network.Mock.Chain (Chain (..))
 import Ouroboros.Network.Mock.Chain qualified as Chain
 import Ouroboros.Network.Mock.ConcreteBlock (Block)
 import Ouroboros.Network.Mock.ConcreteBlock qualified as ConcreteBlock
 import Ouroboros.Network.Mock.ProducerState
+
+import Simulation.Network.Snocket (AddressType (..), GlobalAddressScheme (..))
+
+import Control.Concurrent.Class.MonadMVar.Strict qualified as Strict
 import Ouroboros.Network.NodeToNode ()
 import Ouroboros.Network.NodeToNode.Version (DiffusionMode (..))
 import Ouroboros.Network.PeerSelection (PeerSharing, RelayAccessPoint (..))
@@ -76,12 +84,13 @@ import Ouroboros.Network.PeerSharing (PeerSharingAPI, PeerSharingRegistry (..),
            ps_POLICY_PEER_SHARE_MAX_PEERS, ps_POLICY_PEER_SHARE_STICKY_TIME)
 import Ouroboros.Network.Protocol.Handshake.Unversioned
 import Ouroboros.Network.Snocket (TestAddress (..))
-
-import Simulation.Network.Snocket (AddressType (..), GlobalAddressScheme (..))
-
+import Ouroboros.Network.TxSubmission.Inbound.V2.Registry (SharedTxStateVar,
+           TxChannels (..), TxChannelsVar, TxMempoolSem, newSharedTxStateVar,
+           newTxMempoolSem)
 import Test.Ouroboros.Network.Diffusion.Node.ChainDB (ChainDB (..))
 import Test.Ouroboros.Network.Diffusion.Node.ChainDB qualified as ChainDB
 import Test.Ouroboros.Network.Orphans ()
+import Test.Ouroboros.Network.TxSubmission.Types (Mempool, Tx, newMempool)
 import Test.QuickCheck (Arbitrary (..), choose, chooseInt, frequency, oneof)
 
 
@@ -117,6 +126,11 @@ instance Show NtNAddr_ where
     show (EphemeralIPv6Addr n) = "EphemeralIPv6Addr " ++ show n
     show (IPAddr ip port)      = "IPAddr (read \"" ++ show ip ++ "\") " ++ show port
 
+ppNtNAddr_ :: NtNAddr_ -> String
+ppNtNAddr_ (EphemeralIPv4Addr n) = "eph.v4." ++ show n
+ppNtNAddr_ (EphemeralIPv6Addr n) = "eph.v6." ++ show n
+ppNtNAddr_ (IPAddr ip port)      = show ip ++ ":" ++ show port
+
 instance GlobalAddressScheme NtNAddr_ where
     getAddressType (TestAddress addr) =
       case addr of
@@ -136,6 +150,13 @@ data NtNVersionData = NtNVersionData
   , ntnPeerSharing   :: PeerSharing
   }
   deriving Show
+
+ppNtNAddr :: NtNAddr -> String
+ppNtNAddr (TestAddress addr) = ppNtNAddr_ addr
+
+ppNtNConnId :: ConnectionId NtNAddr -> String
+ppNtNConnId ConnectionId { localAddress, remoteAddress } =
+  ppNtNAddr localAddress ++ "→" ++ ppNtNAddr remoteAddress
 
 instance Acceptable NtNVersionData where
   acceptableVersion
@@ -263,7 +284,7 @@ randomBlockGenerationArgs bgaSlotDuration bgaSeed quota =
     , bgaSeed
     }
 
-data NodeKernel header block s m = NodeKernel {
+data NodeKernel header block s txid m = NodeKernel {
       -- | upstream chains
       nkClientChains
         :: StrictTVar m (Map NtNAddr (StrictTVar m (Chain header))),
@@ -280,12 +301,27 @@ data NodeKernel header block s m = NodeKernel {
 
       nkPeerSharingAPI :: PeerSharingAPI NtNAddr s m,
 
-      nkPublicPeerSelectionVar :: StrictTVar m (PublicPeerSelectionState NtNAddr)
+      nkPublicPeerSelectionVar :: StrictTVar m (PublicPeerSelectionState NtNAddr),
+
+      nkMempool :: Mempool m txid,
+
+      nkTxChannelsVar :: TxChannelsVar m NtNAddr txid (Tx txid),
+
+      nkTxMempoolSem :: TxMempoolSem m,
+
+      nkSharedTxStateVar :: SharedTxStateVar m NtNAddr txid (Tx txid)
     }
 
-newNodeKernel :: MonadSTM m
-              => s -> m (NodeKernel header block s m)
-newNodeKernel rng = do
+newNodeKernel :: ( MonadSTM m
+                 , Strict.MonadMVar m
+                 , SplitGen s
+                 , Eq txid
+                 )
+              => s
+              -> Int
+              -> [Tx txid]
+              -> m (NodeKernel header block s txid m)
+newNodeKernel psRng txSeed txs = do
     publicStateVar <- makePublicPeerSelectionStateVar
     NodeKernel
       <$> newTVarIO Map.empty
@@ -293,15 +329,19 @@ newNodeKernel rng = do
       <*> newFetchClientRegistry
       <*> newPeerSharingRegistry
       <*> ChainDB.newChainDB
-      <*> newPeerSharingAPI publicStateVar rng
+      <*> newPeerSharingAPI publicStateVar psRng
                             ps_POLICY_PEER_SHARE_STICKY_TIME
                             ps_POLICY_PEER_SHARE_MAX_PEERS
       <*> pure publicStateVar
+      <*> newMempool txs
+      <*> Strict.newMVar (TxChannels Map.empty)
+      <*> newTxMempoolSem
+      <*> newSharedTxStateVar (mkStdGen txSeed)
 
 -- | Register a new upstream chain-sync client.
 --
 registerClientChains :: MonadSTM m
-                     => NodeKernel header block s m
+                     => NodeKernel header block s txid m
                      -> NtNAddr
                      -> m (StrictTVar m (Chain header))
 registerClientChains NodeKernel { nkClientChains } peerAddr = atomically $ do
@@ -313,7 +353,7 @@ registerClientChains NodeKernel { nkClientChains } peerAddr = atomically $ do
 -- | Unregister an upstream chain-sync client.
 --
 unregisterClientChains :: MonadSTM m
-                       => NodeKernel header block s m
+                       => NodeKernel header block s txid m
                        -> NtNAddr
                        -> m ()
 unregisterClientChains NodeKernel { nkClientChains } peerAddr = atomically $
@@ -365,34 +405,43 @@ instance Exception NodeKernelError where
 -- | Run chain selection \/ block production thread.
 --
 withNodeKernelThread
-  :: forall block header m seed a.
+  :: forall block header m seed txid a.
      ( Alternative (STM m)
      , MonadAsync         m
      , MonadDelay         m
+     , MonadFork          m
      , MonadThrow         m
      , MonadThrow    (STM m)
+     , Strict.MonadMVar   m
      , HasFullHeader block
      , SplitGen seed
+     , Eq txid
      )
-  => BlockGeneratorArgs block seed
-  -> (NodeKernel header block seed m -> Async m Void -> m a)
+  => NtNAddr
+  -- ^ just for naming a thread
+  -> BlockGeneratorArgs block seed
+  -> [Tx txid]
+  -> (NodeKernel header block seed txid m -> Async m Void -> m a)
   -- ^ The continuation which has a handle to the chain selection \/ block
   -- production thread.  The thread might throw an exception.
   -> m a
-withNodeKernelThread BlockGeneratorArgs { bgaSlotDuration, bgaBlockGenerator, bgaSeed }
+withNodeKernelThread addr BlockGeneratorArgs { bgaSlotDuration, bgaBlockGenerator, bgaSeed }
+                     txs
                      k = do
-    kernel <- newNodeKernel psSeed
+    kernel <- newNodeKernel psSeed txSeed txs
     withSlotTime bgaSlotDuration $ \waitForSlot ->
       withAsync (blockProducerThread kernel waitForSlot) (k kernel)
   where
-    (bpSeed, psSeed) = splitGen bgaSeed
+    (bpSeed, rng) = splitGen bgaSeed
+    (txSeed, psSeed) = random rng
 
-    blockProducerThread :: NodeKernel header block seed m
+    blockProducerThread :: NodeKernel header block seed txid m
                         -> (SlotNo -> STM m SlotNo)
                         -> m Void
     blockProducerThread NodeKernel { nkChainProducerState, nkChainDB }
                         waitForSlot
-                      = loop (Block.SlotNo 1) bpSeed
+                      =  labelThisThread ("krnl-" ++ ppNtNAddr addr)
+                      >> loop (Block.SlotNo 1) bpSeed
       where
         loop :: SlotNo -> seed -> m Void
         loop nextSlot seed = do